+++ src/hooks/queries/useQuestion.ts
@@ -1,4 +1,4 @@
-import { get } from '@/libs/api';
+import { get } from '@/libs/clientSideApi';
 import { useQuery } from '@tanstack/react-query';
 
 type Question = {
